+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/bean/entity/JSFixedFastJsonPexp.java
@@ -16,7 +16,7 @@ import java.util.Objects;
  */
 public class JSFixedFastJsonPexp implements Bean {
 
-    private static final long serialVersionUID = 6913527582734758032L;
+    private static final long serialVersionUID = 4926328534184523648L;
 
     public static JSFixedFastJsonPexp of(Pexp pexp) {
         if (Objects.isNull(pexp)) {
@@ -26,26 +26,37 @@ public class JSFixedFastJsonPexp implements Bean {
                 JSFixedFastJsonLongIdKey.of(pexp.getKey()),
                 FastJsonStringIdKey.of(pexp.getRoleKey()),
                 pexp.getContent(),
+                pexp.getDescription(),
                 pexp.getRemark()
         );
     }
 
     @JSONField(name = "key", ordinal = 1)
     private JSFixedFastJsonLongIdKey key;
+
     @JSONField(name = "role_key", ordinal = 2)
     private FastJsonStringIdKey roleKey;
+
     @JSONField(name = "content", ordinal = 3)
     private String content;
-    @JSONField(name = "remark", ordinal = 4)
+
+    @JSONField(name = "description", ordinal = 4)
+    private String description;
+
+    @JSONField(name = "remark", ordinal = 5)
     private String remark;
 
     public JSFixedFastJsonPexp() {
     }
 
-    public JSFixedFastJsonPexp(JSFixedFastJsonLongIdKey key, FastJsonStringIdKey roleKey, String content, String remark) {
+    public JSFixedFastJsonPexp(
+            JSFixedFastJsonLongIdKey key, FastJsonStringIdKey roleKey, String content, String description,
+            String remark
+    ) {
         this.key = key;
         this.roleKey = roleKey;
         this.content = content;
+        this.description = description;
         this.remark = remark;
     }
 
@@ -73,6 +84,14 @@ public class JSFixedFastJsonPexp implements Bean {
         this.content = content;
     }
 
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
     public String getRemark() {
         return remark;
     }
@@ -83,10 +102,11 @@ public class JSFixedFastJsonPexp implements Bean {
 
     @Override
     public String toString() {
-        return "JsonPexp{" +
+        return "JSFixedFastJsonPexp{" +
                 "key=" + key +
                 ", roleKey=" + roleKey +
                 ", content='" + content + '\'' +
+                ", description='" + description + '\'' +
                 ", remark='" + remark + '\'' +
                 '}';
     }